[问题/需求描述]:适配最新的交易接口 [解决方法]:适配最新的交易接口 [影响范围]:无 [需求编号]:无

+++ domain/components/instrument_info.h
@@ -13,7 +13,7 @@ struct InstrumentInfo {
     int32_t min_market_order_volume;
     int32_t max_limit_order_volume;
     int32_t min_limit_order_volume;
-    int32_t tradeuint;
+    double tradeuint;
     int32_t is_trading;
   };
